feat: Remove references to sunset chinese website (#70)
The chinese website remains almost unmodified since 2019 and there are
ongoing efforts to detach it from our resources.
I would like to remove it from the navbar component. Not only for the
reason that it's in a sunset state, but also to draw focus on jenkins.io
and prevent future misguidance of newbie contributors and GSoC students
who want to contribute.